Cleaned up TestReportStream; created new dbsubmission-load test
